auctions: fix leftover float type hints from the C4 Decimal migration
Pyright flagged 4 type errors that the earlier Float -> Decimal pass
missed:

- InvalidAuctionStartPriceException/InvalidAuctionminimumIncrementException
  still typed their constructor param as float; auction_aggregate.py
  raises them with the Decimal start_price/minimum_increment it now
  holds, so "Decimal is not assignable to float".
- test_auction_application.py built a CreateAuctionSchema with bare
  float literals (100.0, 10.0); the schema field is Decimal, so
  "float is not assignable to Decimal".

Both fixed to use Decimal consistently. No behavior change, pure
typing cleanup.

97/97 tests passing.

+++ app/modules/auction/domain/exceptions/auction_exceptions.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from app.core.exceptions.exceptions import ConflictException, DomainException, NotFoundException
 
 
@@ -17,7 +19,7 @@ class InvalidAuctionDescriptionException(DomainException):
 
 
 class InvalidAuctionStartPriceException(DomainException):
-    def __init__(self, start_price: float):
+    def __init__(self, start_price: Decimal):
         super().__init__(f"Invalid auction start price: {start_price}")
 
 
@@ -27,7 +29,7 @@ class InvalidAuctionEndTimeException(DomainException):
 
 
 class InvalidAuctionminimumIncrementException(DomainException):
-    def __init__(self, minimum_increment: float):
+    def __init__(self, minimum_increment: Decimal):
         super().__init__(f"Invalid auction minimum increment: {minimum_increment}")
